Completed symmetric_closure
symmetric_closure is passing all tests. symmetric_closure adds only the edges necessary to make the relation symmetric. It does this by transposing a copy of itself and then joining itself to the trasnposed copy. The result of this opperation is a symmetric Matrix with only the necessary edges added.